Fix bug in spatial, and add "reverse strategy" functionality to strat_test.
"Reverse strategy" allows for a test to be reversed so that MAB algorithms
find the worst performers rather than the best ones (used to determine if
passing some property automatically makes the method resistant to strategy).

Also fix strange parameter choice in strat_test.h and update test_strategy
files that use this so that they still work. Simplify some of the
test_strategy files too.