Add reverse perspective option to minmax (for debugging).
Reverse perspective mostly makes sense for pairwise opposition, and
instead of scoring candidate A as the negative of the greatest victory
someone has against him, scores him as the minimum strength of any victory
he has against someone else.

With pairwise opposition, this will always produce the same ranking as
ordinary minmax on the same ballot set (also with pairwise opposition),
but without negative values for the scores.