Add invariance property tester.
These test that the outcome stays the same if we multiply every ballot weight
by the same constant, and if we add a constant number of ballots of every
preference.

Properly speaking, these are relative properties and should be handled as
such, but they're not tied to any scenario and so there's a separate program
for them so far. Fix later.