More work on ISDA. Very rough, will redo/incorporate as it gets better.
The problem lies in that ISDA is essentially a composition, but we don't
have direct access to the constraint set generated by test_generator. So
I put the composition inside test_generator, but that makes it ugly because
the rest of the program assumes that the scenarios the test generator will
use are the scenarios passed to it.

Furthermore, when deciding what pairs of candidate indices for B and B' in
test instance generator, the internal relative constraint doesn't know about
the ISDA wrapper/compositor, and so would get confused if we ask it directly
based on the indices outside of the ISDA wrapper. So that has to be handled
some way, too. I think my solution will be useful for non-ISDA as well
(tidying up the process of finding what B, B' pairs are appropriate for e.g.
independence of clone checks), but it needs to be less messy.