brute_rpn: Start adding support for "nonexistent candidates".
These are stand-ins for candidates who are added or eliminated as part of
a relative criterion transformation, but are not derived from candidates
who already exist.

This is required for e.g. ISDA with mono-add-top, where adding some A-top
ballots can push a candidate who wasn't in the Smith set onto that set.