Test_results: make testing more discerning to avoid pathological cases.
These pathological cases include ones that produce a tie among all
candidates no matter what. The more discerning test doesn't just flag
on (for no-harm) A>B before and A'<B' after, but A>=B before and A'<B'
after as well as A>B before and A'=B' after.