brute_rpn: Add support for "nonexistent candidates".
The implementation in polytope_compositor.cc is a bit ugly and could probably
be made quicker. I'll do that if required; but nonexistent candidates at least
works properly now, so ISDA should test for situations where a candidate
enters or leaves the Smith set, not just failures among candidates that stay
in the Smith set both before and after.