Add Strong Unmanipulable Majority criterion to tests.
This is Benham's suggestion on the Electorama list in November 2008, URL
http://lists.electorama.com/pipermail/election-methods-electorama.com/2008-November/088754.html

The "Strong" part refers to that the method fails the criterion if there
exists some pair A and B where the method ranks A ahead of B, but
manipulation can get B ahead of A. Unlike Benham's suggestion, A doesn't
have to be a winner in the pre-manipulation election, but only has to
be ranked ahead of B by the method.

I intended the test to cut down on the number of fpA-fpC extensions onto
four candidates that I would have to test later, since presumably any
method passing SUM would be much harder to manipulate than methods not
passing it, all other things equal. But it seems the SUM is too strong
when used with Benham's original suggestion for Unmanipulable Majority.