Do a light refactor of StrategyTest.
Also add support for distinguishing burial, compromise, two-sided
strategies and everything else.

I ultimately want to merge this with monotonicity and similar tests,
so that the user can just specify some criterion (even if that is
"strategy-proofness") and get information about whether the method
passes the criterion, and if not, how often it is failed under a given
ballot model.

Known regression: The Gaussian vote distribution no longer works with
the strategy generator. (I'll have to find out why.)