strategy_test: Handle multi- and single instance tests generically.
Multi instance tests are like coalitional strategy: there are many
possible strategies. Single instance is like (maximum) burial or
compromising: everybody either puts the compromise candidate top
or the buried candidate bottom.

This code has a very ugly hack. It must be fixed and this is not
appropriate for the long term. TODO.

I also need to check if there's a regression on fixed_strategy's
Gaussian error results.