Fix DSC bug that would lead to incorrect full rankings.
Also improve handling of eliminated candidates in the same method used
in conjunction with loser elimination or average-loser elimination.

The full ranking works by checking which candidate of the remaining
ones can win when intersecting coalitions like DSC; this candidate
is then removed and the process repeated, giving a social ordering from
best to worst.

However, to break ties properly, the method needs to sort the coalitions
so that among coalitions with the same support, those containing the
candidate being tested appear first. This "priority candidate" sorting
tiebreak was handled incorrectly; the priority candidate of the ith
coalition was set to i instead of the actual candidate number to be
tested.