Yee Quasi Monte-Carlo: Disable autopilot.
That's what was interfering with the result precision, not the
dimension problem mentioned in the previous commit.

I tried to expand the number of dimensions as suggested in that commit,
but it just made the convergence worse and introduces non-uniform
glitches, so disregard what I said there about the number of
dimensions required for QMC Yee integration.