Refactor random and quasirandom generators to use a common interface.
It's a bit of a hack, all the business about start_query()/end_query()
that hasn't been implemented yet, but it should provide a general way
to handle both QMC and MC.

The ballot generators still directly refer to rng, not coordinate_gen;
that'll have to be fixed later as I update them to use start/end query
or get_coordinates.

There are also some numerical precision things that need to be evened
out in the QMC class.