Make ballot generators use coord generators instead of RNGs directly.
Also extract some Gaussian distribution code from the Gaussian and
Quasi-Gaussian ballot generators to its own code. This makes the two
generators nearly identical (they just reference different coord
generators).

But passing the right coord generators is going to be difficult. For
instance, for Yee, we want the candidate positions to be chosen randomly,
not quasi-randomly, even when doing quasi-random sampling for the diagram
itself.

And very few of the other ballot generators will work with a QMC source
yet, because they don't use start_query/end_query yet and the dimensions
aren't properly set. All of that needs to be done later.