Refactor: make modes use coordinate generators instead of RNGs too.
This refactor is rather ugly, see the TODOs and HACKs, but it's a
start, and this way of doing it is really better than special-casing
Quasi-Monte Carlo and using mutable objects.

To do this properly, I'll probably have to distinguish the different
"features" that (quasi) random numbers are used for: e.g. in Bayesian
Regret, we draw from a 2D space to determine the number of voters and
candidates, then from a larger space to sample these voters' ballots.
What makes it tougher is that e.g. impartial culture's dimensions
depend on the number of voters and candidates.