Refactor: allow modes to use multiple coordinate sources.
This is done by calling set_coordinate_gen with a valid purpose
(what the coordinate generator is going to be used for) and a
shared pointer to the generator. The setup functions in
src/main/combined make use of these calls to set the appropriate
generators.

As a bonus, the completely deterministic modes (e.g. the ballot
interpreter) no longer needs a coordinate generator parameter passed
to their init functions.