Define election_t, switch from std::set to boost flat_set for ~1.8x speedup.
A hand-made "frozen election" read-only/lazily sorted structure could
probably be much faster, but this is almost a free speedup, so why not?

The change from std::list<ballot_group> to election_t will make it easier
to change things later.

Also do a few optimizations to spatial_generator like not declaring the
ordering structure inside the inner loop. This cuts down on a few deletes
and mallocs, although there seems to be something off with copy elision
not working and leading to a bunch of unnecessary memory allocates and
deletions, which is slowing down the Yee process compared to the reference
benchmark.