Add Disqualification-Elimination code.
The method is described in comments, but see also
http://lists.electorama.com/pipermail/election-methods-electorama.com/2024-February/005353.html
for more information.

I haven't added it to get_methods yet because if I don't use a base
method, the method will violate neutrality, and I don't want that for
methods exposed through the get_methods functions. On the other hand,
if I do use a base method, then which should it be?