Refactor: Turn "tests" into "simulators", detach from bandit stuff.
This will make it easier to write new simulators like VSE or computational
complexity; and to slot these into multi-armed bandit as needed.

The lil'UCB class still needs a bunch of terminology cleanup, but at least
it works. In addition, I need to set some simulators to indicate that
lower is better (e.g. lower manipulability is better) so that I can remove
the ugly kludges involving inverting the return values in the bandit
manipulability checks.

The confidence interval code has also been disabled for now, and needs to be
implemented properly again with special casing for Bernoulli (0/1)
simulators.