Add cardinal methods to strategy test and ballot operator== HACK. SLOW!
The hack now considers orderings with equal rank but different scores
to be different. The candscore class doesn't know whether the ballot data
it represents is part of a rated or ranked ballot, so it can't meaningfully
disregard rated info (different scores, same rank) only when dealing with
ranked ballots.

Prior versions used to assume every ballot was ranked, which led ballots
with different ratings but the same rank to be compressed together when
ballot compression is enabled, thus leading to weird results with rated
methods. Now it assumes every ballot is rated, which works, but is *much*
slower.

This is an imperfect solution to getting rated methods working. I need
to figure out a better way of doing it, and signs point to having to
redo the ballot data structure somehow.