Make monotonicity and strat checker Bernoulli, improve bandit and test code.
This reclasses the monotonicity and strategy tests to be Bernoulli (since
they return either failure of a criterion or no failure detected). It also
adds functionality to pull bandit arms for a given period of time rather
than a given number of iterations, and cleans up test_monotonicity_bandit
to use confidence intervals and timed bandit pulls.