Add preliminary VSE test program.
What this shows so far is that, while the VSE calculation does seem to
work, it's very sensitive to imprecision in the initial calculation of
E[optimal] - E[random]. If I can get this analytically - the true limit
at infinity - then there would be no problem, but I can't at this point.

So while the logic that suggests that VSE can be linearized and used in
a bandit setting is sound, I might need some way to show the "finite" VSE
to the user, so that when the number of iterations exceeds the number
used to approximate E[optimal] - E[random] at the start, the value will
keep converging to the proper value at infinity instead of always having
some finite imprecision. How to present that, coding wise, could be
somewhat difficult, though...