Fix MSVC compile problems. Includes HACK: commenting out mmap functionality.
The functions using mmap now throw an exception instead on Windows.
This needs to be replaced with a proper independent mmap implementation
or the code has to be rewritten somehow.

Other compile problems that have been fixed are the (extraneous) use of
values.h, a reference to unistd.h, and dynamic allocation of a char[],
as well as more uses of the MSVC random hack.

Also fix an uninitialized parameter bug.