Add multiwinner code from an earlier project from 2008.
This code shares a lot of common stuff with quadelect, but can't be
directly included without some serious adaptation. I've thus not added
it to CMakeLists yet; better to first commit the source itself, and only
then start working on it.

multiwinner.cc calculates proportionality and majoritarian quality
(related to VSE). See
http://lists.electorama.com/pipermail/election-methods-electorama.com/2008-June/087136.html
for more information.