Add combinations library/code, update QPQ, and reorganize a bit.
The combinations code is from
https://github.com/HowardHinnant/combinations
and is intended to be used to simplify hard_card and similar combinatorial
multiwinner methods later.

The QPQ code is from a more recent version of my earlier code, but still
not ported.

And SpookyHash is now in the src/lib directory so that all external code is
in one place.