Implement Zbornik's isoelastic/constant risk suggestion; normalize it and LPV.
My earlier multiwinner-only program also implemented this method. His
suggestion is for approval alone; I've extended it to range in a
straightforward way.

Normalizing the ratings seems to improve the performance of both
isoelastic and LPV, so add that too. It should ideally be a parameter, or
the ratings should be transformed before they're given to the methods.