Fix unused variables and variable with a undefined value.
The latter isn't actually used anywhere before it's declared, and
an assert would enforce this, but it would still trigger a warning.
Now it's set to a value that's never used, and the assert has been
replaced with an exception so it can't be used in a release build
either. (optima.h)