Remove the complex tiebreaker stuff from QPQ.
This will fail neutrality if there are two candidates that are equally
"worst" or "best" (for elimination or election respectively). But this
should be pretty easy to check with has_multiple_winners/losers. Maybe
do that later.

We lose a little bit of utilitarian VSE and gain a little bit of
proportionality by doing this. Using a Condorcet tiebreaker above or
below the quota is also an option; I don't think it would change the
figures much, but it could be used to give an incentive for broader
appeal. (See Ogren and Robinette's papers.)