Adjust methods' proportionality VSE to be relative to random council.
Not random dictator, as it used to be; this will make it easier to
compare the Droop region plot to the methods' position.

(However, the region plot may still be buggy. PSC-CLE seems to be outside
the region despite passing Droop. More research is required.)

Also add some text distinguishing harmonic voting with delta > 1 from
0 <= delta <= 1, since the former is possible (doesn't blow the method
up) yet out of Warren's spec.