Multiwinner: fix Sainte-Laguë index bug, add timed output, extract convex_hull.
The main multiwinner program was passing "observed" and "expected" parameters
(opinion profile of the elected candidates and the voters respectively) in the
wrong order. I didn't spot this because when I started programming the
code in 2008, I used the RMS error, where the order of these parameters
don't matter. But for the Sainte-Laguë index, they do.
(This still doesn't entirely resolve the oddity Toby Pereira informed me
of. More investigation is needed.)

The multiwinner program now also only dumps method results ever so often,
so that it doesn't emit lots of info we don't need anyway.

And most of the convex hull code is now in its own .cc file.