Cycle-cutting: add parameter for pairwise type.
This should make it easier to figure out what's going on, but I still
need to do the actual debugging. Preliminary strategy results suggest
that margins is best for impartial culture, and fwv is best for the 4D
spatial model, but margins isn't too bad either.

But it would still be better to figure out what's going on.