Add "hash-random" method.
This method takes all the ballot data and runs it through a hash
function, ordering candidates according to the hash value when
setting the first double of the data to the candidate number. In
essence, it's a reproducible version of random candidate, and as
such extremely manipulable.

It's useful for finding bounds on manipulability of certain sets
or other methods over all possible ways to break ties.